+++ Source/Coinbase.Commerce/Models/JsonResponse.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
@@ -21,8 +22,30 @@ namespace Coinbase.Commerce.Models
       [JsonProperty("error")]
       public Error Error { get; set; }
 
+      /// <summary>
+      /// Responses can include a warnings parameter to notify the developer
+      /// of best practices, implementation suggestions or deprecation warnings.
+      /// While you don’t need show warnings to the user, they are usually
+      /// something you need to act on.
+      /// </summary>
       [JsonProperty("warnings")]
       public string[] Warnings { get; set; }
+
+      /// <summary>
+      /// Checks if the response has errors.
+      /// </summary>
+      public bool HasError()
+      {
+         return this.Error != null;
+      }
+
+      /// <summary>
+      /// Checks if the response has warnings.
+      /// </summary>
+      public bool HasWarnings()
+      {
+         return this.Warnings?.Length > 0;
+      }
    }
 
    public class Response<T> : JsonResponse