+++ Source/Coinbase.Tests/Examples.cs
@@ -187,5 +187,8 @@ namespace Coinbase.Tests
                         }
                   })
             };
+
+      public const string WebhookHeaderSignature = "e337ef2d998015288b742d2d43024088fedf8333dc55043a338fd58eff1b282b";
+      public const string Webhook = @"{""attempt_number"":1,""event"":{""api_version"":""2018-03-22"",""created_at"":""2018-04-04T23:49:00Z"",""data"":{""code"":""8EKMDPVQ"",""name"":""Candy Bar"",""pricing"":{""local"":{""amount"":""1.00"",""currency"":""USD""},""bitcoin"":{""amount"":""0.00014691"",""currency"":""BTC""},""ethereum"":{""amount"":""0.002632000"",""currency"":""ETH""},""litecoin"":{""amount"":""0.00841928"",""currency"":""LTC""},""bitcoincash"":{""amount"":""0.00153559"",""currency"":""BCH""}},""metadata"":{""customerId"":""ccaf106d-7fd1-4b8e-b409-0fa345e4d82b""},""payments"":[],""timeline"":[{""time"":""2018-04-04T23:49:00Z"",""status"":""NEW""}],""addresses"":{""bitcoin"":""1Fir6DxwEbQdLWeDMYrqgpnYrs2mDLxobc"",""ethereum"":""0x116462dd6260ecc1b10a79f07bb30af10affc886"",""litecoin"":""Lh8He9HadDszxmY33GTngqQYLDSHaPRLaW"",""bitcoincash"":""qqs2llv7yz76d7c5kkjnpazzllwq7sq29546zt5edc""},""created_at"":""2018-04-04T23:49:00Z"",""expires_at"":""2018-04-05T00:04:00Z"",""hosted_url"":""https://commerce.coinbase.com/charges/8EKMDPVQ"",""description"":""Sweet Tasting Chocolate"",""pricing_type"":""fixed_price""},""id"":""f6972c57-c100-4e64-b47c-193adecfadc6"",""type"":""charge:created""},""id"":""0d0cee27-8ffb-405d-9d05-3326dd4bcb7d"",""scheduled_for"":""2018-04-04T23:49:00Z""}";
    }
 }