+++ HISTORY.md
@@ -1,2 +1,7 @@
+# v1.0.1
+* Added `.HasError()` and `.HasWarnings()` helper methods on JSON response objects.
+* Set `X-CC-Version` header to `2018-03-22` to remove server warnings.
+* Set `User-Agent` string when making API calls.
+
 # v1.0.0
 * Initial release.